+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 * Feature
   - Add: `StkTableColumn['sortConfig']` to support column localeCompare
   - `StkTableColumn['minWidth']` Can affect the minimum width of column width adjustment.
+* Optimize
+  - `StkTableColumn['width']` if set number string, auto append `px`.
 
 ## 0.6.3
 * Change: old key `dataIndex` replace to `props.colKey` result.